core: Move `content_hash` and `repo_path` into it
This part of the library is quite basic and it doesn't have too many dependencies which also need to 
move to make it happen. We also need to move the `ContentHash` macro since all `RepoPath` types depend on it,
the macro is moved to a new `jj-core-proc-macros` crate which mirrors the existing structure in `jj-lib`.
Since we now have `jj-core-proc-macros` this also deprecates `jj-lib-proc-macros` for external consumers.

Part of #6284
