+++ Cargo.toml
@@ -139,6 +139,8 @@ winreg = "0.56"
 
 # put all inter-workspace libraries, i.e. those that use 'path = ...' here in
 # their own (alphabetically sorted) block
+jj-core = { path = "lib/core", version = "0.43.0", default-features = false }
+jj-core-proc-macros = { path = "lib/core/proc-macros", version = "0.43.0" }
 jj-lib = { path = "lib", version = "0.43.0", default-features = false }
 jj-lib-proc-macros = { path = "lib/proc-macros", version = "0.43.0" }
 testutils = { path = "lib/testutils" }
