+++ lib/Cargo.toml
@@ -47,7 +47,8 @@ hashbrown = { workspace = true }
 indexmap = { workspace = true }
 interim = { workspace = true }
 itertools = { workspace = true }
-jj-lib-proc-macros = { workspace = true }
+jj-core = { workspace = true }
+jj-core-proc-macros = { workspace = true }
 maplit = { workspace = true }
 memchr = { workspace = true }
 once_cell = { workspace = true }
