+++ lib/core/proc-macros/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "jj-core-proc-macros"
+description = "Proc macros for the jj-core crate"
+
+categories.workspace = true
+documentation.workspace = true
+edition.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+license.workspace = true
+readme.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+include = ["/LICENSE", "/src/"]
+
+[lib]
+proc-macro = true
+
+[dependencies]
+proc-macro2 = { workspace = true }
+quote = { workspace = true }
+syn = { workspace = true }
+
+[lints]
+workspace = true
